feat: Add search bar to patient home screen
A `SearchView` has been added to the `fragment_patient_home_page.xml` layout, allowing users to search for services.

To style the new search bar, a `MaterialCardView` is used to provide elevation and rounded corners. The background of the `SearchView` is set to white. The query hint is set to "Search for MediTrack Services".

Additionally, some attributes in the `fragment_patient_home_page.xml` file have been reordered for better readability.