Refactor: Improve patient home page UI and logic
Refactor the initialization of UI components in `PatientHomePageFragment` by moving them into the methods where they are used (`checkInvites` and `getMeds`).

Add a new `TextView` (`noMedToday`) to the layout, which will be displayed when there are no upcoming medications scheduled for the current day. Update the string resource for "no medicine scheduled" to specify "Today".